refactor(examples): extract thumbnails module
